pyproject.toml
[project]
name = "career-presence"
version = "1.0.0"
description = "Automated job search, resume tailoring, and professional presence management"
readme = "README.md"
license = { text = "PolyForm-Noncommercial-1.0.0" }
requires-python = ">=3.11"
authors = [
    { name = "Career Presence Contributors" }
]
keywords = [
    "job-search",
    "resume",
    "career",
    "automation",
    "linkedin",
    "ats"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # HTTP & Web
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",

    # Browser Automation
    "playwright>=1.41.0",

    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # CLI & Output
    "rich>=13.7.0",
    "typer>=0.9.0",

    # Async Support
    "aiofiles>=23.2.0",
    "anyio>=4.2.0",

    # Templating
    "jinja2>=3.1.0",

    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",

    # Logging
    "structlog>=24.1.0",

    # Job Search
    "python-jobspy>=1.1.80",
]

[project.optional-dependencies]
# AI/LLM integrations
ai = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.6.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
]

# Vector search for semantic matching
vectors = [
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
]

# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.21.0",
]

# All optional dependencies
all = [
    "career-presence[ai,vectors,dev]",
]

[project.scripts]
# CLI entry points
cps = "scripts.cli:app"
ats-score = "scripts.analysis.ats_scorer:main"
tailor = "scripts.tailoring.resume_tailor:main"
discover = "scripts.discovery.job_searcher:main"
track = "scripts.tracking.tracker:main"

[project.urls]
Homepage = "https://github.com/yourusername/career-presence"
Documentation = "https://github.com/yourusername/career-presence#readme"
Repository = "https://github.com/yourusername/career-presence"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

# ═══════════════════════════════════════════════════════════════════════════
# TOOL CONFIGURATIONS
# ═══════════════════════════════════════════════════════════════════════════

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["scripts", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["scripts"]
branch = true
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
